+++ src/index.js
@@ -29,6 +29,7 @@ import normalize from './utils/normalize.js';
 import radians from './utils/radians.js';
 import random from './utils/random.js';
 import randomInt from './utils/randomInt.js';
+import smoothstep from './utils/smoothstep.js';
 import step from './utils/step.js';
 
 import Vec2 from './vectors/Vec2.js';
@@ -67,6 +68,7 @@ export default {
   radians,
   random,
   randomInt,
+  smoothstep,
   step,
 
   Vec2,